requirements.txt
# Core
fastapi
uvicorn
pydantic
streamlit

# ML & RAG
langchain
openai
faiss-cpu
chromadb
sentence-transformers

# Utils
python-dotenv
pandas
PyPDF2
python-multipart
pypdf
PyMuPDF